Abstract behaviors.
Previous behavior implementation hard coded behaviors/labels/paths into options page and into `enqueue_js`.

Instead, abstract out the behaviors into a variable within the class, loop on options page, and kick as early as possible on `enqueue` if no behaviors are loaded.

This patch also uses the core `selected()` API on the options screen, and prevents an error for treating an uninitialized variable as an array within `enqueue_js`.

This way, if another behavior is added, it is simply a matter of adding a behavior to the array at the top of the main class.

No changes necessary to the database.